overflow fixed

+++ lcd_canvas.c
@@ -171,7 +171,7 @@ void LCD_canvas_putcol(uint16_t color) {
 }
 
 uint8_t LCD_canvas_get_y_overflow() {
-  if (canvas_y + (canvas_pos / (canvas_width + 1)) > canvas_hardY2) {
+  if (canvas_y + ((int32_t)canvas_pos / (canvas_width + 1)) > (int32_t)canvas_hardY2) {
     return 1;
   } else {
     return 0;
